+++ app/src/main/java/com/zxltrxn/workulator/domain/repositoryinterfaces/TaskRepository.kt
@@ -7,5 +7,5 @@ import kotlinx.coroutines.flow.StateFlow
 interface TaskRepository {
     fun createTask(task:TaskModel)
     fun updateTask(newTask:TaskModel)
-    fun getTaskIds(): StateFlow<List<UInt>>
+    fun getTaskIds(): StateFlow<List<Long>>
 }
